+++ src/Async/Resumable.php
@@ -1,0 +1,42 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\TimeWarp\Async;
+
+use Innmind\Immutable\{
+    Attempt,
+    SideEffect,
+};
+
+/**
+ * @internal
+ * @psalm-immutable
+ */
+final class Resumable
+{
+    /**
+     * @param Attempt<SideEffect> $result
+     */
+    private function __construct(
+        private Attempt $result,
+    ) {
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @param Attempt<SideEffect> $result
+     */
+    public static function of(Attempt $result): self
+    {
+        return new self($result);
+    }
+
+    /**
+     * @return Attempt<SideEffect>
+     */
+    public function unwrap(): Attempt
+    {
+        return $this->result;
+    }
+}